requirements.txt
beautifulsoup4==4.12.3
Pillow==9.5.0
Pillow==10.4.0
Requests==2.32.3
streamlit==1.28.2
